\name{GenomicMating-package}
\alias{GenomicMating-package}
\alias{GenomicMating}
\docType{package}
\title{
\packageTitle{GenomicMating}
}
\description{Implements the approach in the recently 
	published article: Akdemir, Deniz, and 
	Julio I. Sanchez. ''Efficient Breeding by Genomic Mating.'' Frontiers in Genetics 7 (2016).}

\author{
Maintainer: Deniz Akdemir <deniz.akdemir.work@gmail.com>
}

\references{
Akdemir, Deniz, and Julio I. Sanchez. "Efficient Breeding by Genomic Mating." Frontiers in Genetics 7 (2016).
}