\name{GenomicMating-package}
\alias{GenomicMating-package}
\alias{GenomicMating}
\docType{package}
\title{
\packageTitle{GenomicMating}
}
\description{Implements the mate selection approach in: Akdemir and Sanchez. ''Efficient Breeding by Genomic Mating.'' Frontiers in Genetics (2016). Importance parameters were rescaled to the range [0,1]. A function to find points on the frontier was added in V1. V2.0 includes additional methods of calculation of mating statistics such as described in Lehermeier (2018) and using simulation of progeny in addition to optimization over multiple traits.}

\author{
Maintainer: Deniz Akdemir <deniz.akdemir.work@gmail.com>
Contributers: 
}

\references{
Akdemir, Sanchez. "Efficient Breeding by Genomic Mating." Frontiers in Genetics (2016).

Lehermeier at al. "Genetic gain increases by applying the usefulness criterion with improved variance prediction in selection of crosses" Genetics (2017).

Broman et al. "R/qtl: QTL mapping in experimental crosses." Bioinformatics (2003).

VanRaden, Paul M. ''Efficient methods to compute genomic predictions.'' Journal of dairy science (2008).
}