\name{GenomicMating-package}
\alias{GenomicMating-package}
\alias{GenomicMating}
\docType{package}
\title{
\packageTitle{GenomicMating}
}
\description{Implements the approach in: Akdemir, Deniz, and 
	Julio I. Sanchez. ''Efficient Breeding by Genomic Mating.'' Frontiers in Genetics 7 (2016). Importance parameters were rescaled to the range [0,1]. A function to find points on the frontier was added.}

\author{
Maintainer: Deniz Akdemir <deniz.akdemir.work@gmail.com>
}

\references{
Akdemir, Deniz, and Julio I. Sanchez. "Efficient Breeding by Genomic Mating." Frontiers in Genetics 7 (2016).
}